pull down from gnumeric
2005-02-16  Jody Goldberg <jody@gnome.org>

	* utils/go-file.c (go_url_check_extension) : pull down from gnumeric

	* gui-utils/go-gui-utils.c (go_gtk_button_new_with_stock) : renamed
	  from go_gtk_button_new_with_stock_image.
	(go_libglade_new) : use go_sys_data_dir.
	(go_gtk_widget_disable_focus) : renamed from go_widget_disable_focus.
	(go_pango_measure_string) : renamed from go_measure_string.
	(go_gtk_window_set_transient) : renamed from go_window_set_transient.
	(go_gtk_nonmodal_dialog) : renamed from gnumeric_non_modal_dialog.
	(go_gtk_file_sel_dialog) : renamed from gnumeric_dialog_file_selection.
	(go_gtk_select_image) : renamed from gui_image_file_select.
	(gui_get_image_save_info) : pulled in from gnumeric.
	(go_atk_setup_label) : renamed from gnm_setup_label_atk.
	(go_help_display) : pulled down from gnumeric and made generic.
	(go_gtk_help_button_init) : ditto.
	(go_gtk_url_is_writeable) : ditto.
	(go_gtk_dialog_run) : pulled from gnumeric
	(go_gtk_notice_dialog) : ditto.
	(go_gtk_notice_nonmodal_dialog) : ditto.
	(go_gtk_query_yes_no) : ditto.

	* gui-utils/Makefile.am : Add libhtml_stub

	* graph/gog-renderer-pixbuf.c (gog_renderer_pixbuf_init) : init the
	  dpis.
	(gog_renderer_pixbuf_get_pango_context) : use the stored calculated
	  dpi rather than gconf
	(gog_renderer_pixbuf_update) : store the dpi

	* goffice.c (libgoffice_init) : init the paths.  We may want to expand
	  this at some point to allow cmd-line or env-var override.
	(go_sys_data_dir) : new.
	(go_sys_icon_dir) : new.

2005-02-16  Jody Goldberg <jody@gnome.org>

	* src/libgnumeric.c (gnm_pre_parse_init) : set the path for the new
	  gnumeric-functions message catalog

	* src/gnumeric-gconf.c : Add an ifdef to silence the gconf warnings
	* src/io-context.[ch], src/io-context-priv.h : move into goffice/app

+++ plugins/qpro/qpro-read.c
@@ -38,7 +38,7 @@
 #include <style-color.h>
 #include <parse-util.h>
 #include <goffice/app/module-plugin-defs.h>
-#include <io-context.h>
+#include <goffice/app/io-context.h>
 
 #include <gsf/gsf-utils.h>
 #include <gsf/gsf-input.h>
@@ -397,7 +397,7 @@ qpro_parse_formula (QProReadState *state, int col, int row,
 			ref.col_relative = (tmp & 0x4000) ? TRUE : FALSE;
 			ref.row_relative = (tmp & 0x2000) ? TRUE : FALSE;
 			if (ref.row_relative)
-				ref.row = ((gint16)(tmp & 0x1fff) << 3) >> 3;
+				ref.row = (int)(((gint16)((tmp & 0x1fff) << 3)) >> 3);
 			else
 				ref.row = tmp & 0x1fff;
 			expr = gnm_expr_new_cellref (&ref);
@@ -415,7 +415,7 @@ qpro_parse_formula (QProReadState *state, int col, int row,
 			a.col_relative = (tmp & 0x4000) ? TRUE : FALSE;
 			a.row_relative = (tmp & 0x2000) ? TRUE : FALSE;
 			if (a.row_relative)
-				a.row = ((gint16)(tmp & 0x1fff) << 3) >> 3;
+				a.row = (int)(((gint16)((tmp & 0x1fff) << 3)) >> 3);
 			else
 				a.row = tmp & 0x1fff;
 
@@ -425,7 +425,7 @@ qpro_parse_formula (QProReadState *state, int col, int row,
 			b.col_relative = (tmp & 0x4000) ? TRUE : FALSE;
 			b.row_relative = (tmp & 0x2000) ? TRUE : FALSE;
 			if (b.row_relative)
-				b.row = ((gint16)(tmp & 0x1fff) << 3) >> 3;
+				b.row = (int)(((gint16)((tmp & 0x1fff) << 3)) >> 3);
 			else
 				b.row = tmp & 0x1fff;
